Rename hard-boiled-eggs recipe to boiled-eggs
Change name of hard-boiled-eggs.html to boiled-eggs.html in same
directory, update <title> and <h1> in boiled-eggs.html to reflect
new recipe name, and update href and display text to boiled eggs
recipe page accordingly in index.html.

Naming the recipe hard-boiled eggs did not reflect the intention
of the actual recipe, which includes instructions for making eggs in
a range of degrees of doneness.

It should now be more clear to user what the possible outcomes are
from this recipe when navigating to it from home page.